Fix premature "destroy me" on incoming app
Fix an issue where the first message the middleware gets about an app is
the "should_exit" message. In this case, the data has no etype and it's
not registered. Previous behavior was to just blindly shove the data to
the factory, which has no idea what to do with this data. This change
validates the data before passing it to the factory.